Fix chapter jump unlocks
Fixes #48
Note: If applied as an update, this will lock many chapters on the games of people who haven't completed the game due to the way it changes chapter jump IDs.  Anyone applying this as an update needs to make a save before applying the update, and the locked chapters will be re-unlocked upon finishing a chapter after loading the save.