Minor miswording: implicit -> complicit
It wouldn't make sense for the Onigafuchi Guardians to be "implicit" in the kidnapping, unless they were commonly known for kidnapping and assumed to be in on any incidents. This is not the case in the scenario. Rather, there is a possibility that they are "complicit" in the kidnapping, meaning they are potentially involved with the crime.